+++ scaper/core.py
@@ -1688,6 +1688,8 @@ class Scaper(object):
             with _close_temp_files(tmpfiles):
                 isolated_events_audio_path = []
 
+                role_counter = {'background': 0, 'foreground': 0}
+
                 for i, e in enumerate(ann.data):
                     if e.value['role'] == 'background':
                         # Concatenate background if necessary. Right now we
@@ -1827,10 +1829,13 @@ class Scaper(object):
                         else:
                             event_folder = isolated_events_path
 
+                        _role_count = role_counter[e.value['role']]
                         event_audio_path = os.path.join(
                             event_folder, 
                             '{:s}{:d}_{:s}{:s}'.format(
-                                e.value['role'], i, e.value['label'], ext))
+                                e.value['role'], _role_count, e.value['label'], ext))
+                        role_counter[e.value['role']] += 1
+                        
                         if not os.path.exists(event_folder):
                             # In Python 3.2 and above we could do 
                             # os.makedirs(..., exist_ok=True) but we test back to