+++ tests/test_core.py
@@ -1454,14 +1454,40 @@ def _test_generate_isolated_events(SR, isolated_events_path=None, atol=1e-4, rto
         isolated_event_audio_paths = ann.sandbox.scaper.isolated_events_audio_path
         isolated_audio = []
 
+        role_counter = {
+            'background': 0,
+            'foreground': 0
+        }
+
         for event_spec, event_audio_path in zip(ann, isolated_event_audio_paths):
             # event_spec contains the event description, label, etc
             # event_audio contains the path to the actual audio
 
             # make sure the path matches the event description
-            assert event_spec.value['role'] in event_audio_path
-            assert event_spec.value['label'] in event_audio_path
-            isolated_audio.append(soundfile.read(event_audio_path)[0])
+            look_for = '{:s}{:d}_{:s}.wav'.format(
+                event_spec.value['role'], 
+                role_counter[event_spec.value['role']],
+                event_spec.value['label']
+            )
+
+            expected_path = os.path.join(isolated_events_path, look_for)
+            # make sure the path exists
+            assert os.path.exists(expected_path)
+
+            # make sure what's in the sandbox also exists
+            assert os.path.exists(event_audio_path)
+            # is an audio file with the same contents as what we expect
+            _isolated_expected_audio, sr = soundfile.read(expected_path)
+            _isolated_sandbox_audio, sr = soundfile.read(event_audio_path)
+            assert np.allclose(_isolated_sandbox_audio, _isolated_expected_audio)
+
+            # make sure the filename matches
+            assert look_for == os.path.basename(event_audio_path)
+
+            # increment for the next role
+            role_counter[event_spec.value['role']] += 1
+
+            isolated_audio.append(_isolated_sandbox_audio)
 
         # the sum of the isolated audio should sum to the soundscape
         assert np.allclose(sum(isolated_audio), soundscape_audio, atol=1e-4, rtol=1e-8)